Address comments by @minwoox
