+++ grpc-protocol/src/main/java/com/linecorp/armeria/common/grpc/protocol/ArmeriaMessageDeframerHandler.java
@@ -47,19 +47,16 @@
 package com.linecorp.armeria.common.grpc.protocol;
 
 import static com.google.common.base.Preconditions.checkState;
-import static java.util.Objects.requireNonNull;
 
 import java.io.FilterInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.Objects;
 
 import javax.annotation.Nullable;
 
 import com.google.common.annotations.VisibleForTesting;
 
 import com.linecorp.armeria.common.annotation.UnstableApi;
-import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageDeframerHandler.DeframedMessage;
 import com.linecorp.armeria.common.stream.HttpDeframerHandler;
 import com.linecorp.armeria.common.stream.HttpDeframerInput;
 import com.linecorp.armeria.common.stream.HttpDeframerOutput;
@@ -200,91 +197,6 @@ public class ArmeriaMessageDeframerHandler implements HttpDeframerHandler<Defram
                    "Deframing has already started, cannot change decompressor mid-stream.");
         this.decompressor = decompressor;
         return this;
-    }
-
-    /**
-     * A deframed message. For uncompressed messages, we have the entire buffer available and return it
-     * as is in {@code buf} to optimize parsing. For compressed messages, we will parse incrementally
-     * and thus return a {@link InputStream} in {@code stream}.
-     */
-    @UnstableApi
-    public static final class DeframedMessage {
-        private final int type;
-
-        @Nullable
-        private final ByteBuf buf;
-        @Nullable
-        private final InputStream stream;
-
-        /**
-         * Creates a new instance with the specified {@link ByteBuf} and {@code type}.
-         */
-        @VisibleForTesting
-        public DeframedMessage(ByteBuf buf, int type) {
-            this(requireNonNull(buf, "buf"), null, type);
-        }
-
-        /**
-         * Creates a new instance with the specified {@link InputStream} and {@code type}.
-         */
-        @VisibleForTesting
-        public DeframedMessage(InputStream stream, int type) {
-            this(null, requireNonNull(stream, "stream"), type);
-        }
-
-        private DeframedMessage(@Nullable ByteBuf buf, @Nullable InputStream stream, int type) {
-            this.buf = buf;
-            this.stream = stream;
-            this.type = type;
-        }
-
-        /**
-         * Returns the {@link ByteBuf}.
-         *
-         * @return the {@link ByteBuf}, or {@code null} if not created with
-         *         {@link #DeframedMessage(ByteBuf, int)}.
-         */
-        @Nullable
-        public ByteBuf buf() {
-            return buf;
-        }
-
-        /**
-         * Returns the {@link InputStream}.
-         *
-         * @return the {@link InputStream}, or {@code null} if not created with
-         *         {@link #DeframedMessage(InputStream, int)}.
-         */
-        @Nullable
-        public InputStream stream() {
-            return stream;
-        }
-
-        /**
-         * Returns the type.
-         */
-        public int type() {
-            return type;
-        }
-
-        @Override
-        public boolean equals(@Nullable Object o) {
-            if (this == o) {
-                return true;
-            }
-            if (!(o instanceof DeframedMessage)) {
-                return false;
-            }
-
-            final DeframedMessage that = (DeframedMessage) o;
-
-            return type == that.type && Objects.equals(buf, that.buf) && Objects.equals(stream, that.stream);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(buf, stream);
-        }
     }
 
     /**
