+++ grpc/src/main/java/com/linecorp/armeria/internal/client/grpc/InternalGrpcWebUtil.java
@@ -26,7 +26,7 @@ import com.google.common.io.ByteStreams;
 
 import com.linecorp.armeria.common.HttpHeaders;
 import com.linecorp.armeria.common.HttpHeadersBuilder;
-import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageDeframerHandler.DeframedMessage;
+import com.linecorp.armeria.common.grpc.protocol.DeframedMessage;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.ByteBufAllocator;
