+++ grpc/src/main/java/com/linecorp/armeria/server/grpc/ArmeriaServerCall.java
@@ -48,9 +48,9 @@ import com.linecorp.armeria.common.grpc.GrpcJsonMarshaller;
 import com.linecorp.armeria.common.grpc.GrpcSerializationFormats;
 import com.linecorp.armeria.common.grpc.GrpcWebTrailers;
 import com.linecorp.armeria.common.grpc.ThrowableProto;
-import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageDeframerHandler.DeframedMessage;
 import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageFramer;
 import com.linecorp.armeria.common.grpc.protocol.Decompressor;
+import com.linecorp.armeria.common.grpc.protocol.DeframedMessage;
 import com.linecorp.armeria.common.grpc.protocol.GrpcHeaderNames;
 import com.linecorp.armeria.common.logging.RequestLogProperty;
 import com.linecorp.armeria.common.stream.AbortedStreamException;
@@ -553,7 +553,7 @@ final class ArmeriaServerCall<I, O> extends ServerCall<I, O>
 
     private void setClientStreamClosed() {
         if (!clientStreamClosed) {
-            messageDeframer().abort();
+            messageDeframer().close();
             clientStreamClosed = true;
         }
     }
